requirements.txt
fastapi
uvicorn
httpx
python-dotenv
streamlit
sqlite-utils
